+++ ArakaDev/ArakaDev.vcxproj
@@ -13,6 +13,7 @@
   <ItemGroup>
     <ClCompile Include="..\Bullet.cpp" />
     <ClCompile Include="..\Collision.cpp" />
+    <ClCompile Include="..\Dawn.cpp" />
     <ClCompile Include="..\Jilltest.cpp" />
     <ClCompile Include="..\Rendering.cpp" />
   </ItemGroup>
@@ -20,6 +21,7 @@
     <ClInclude Include="..\Bullet.h" />
     <ClInclude Include="..\Collision.h" />
     <ClInclude Include="..\Constants.h" />
+    <ClInclude Include="..\Dawn.h" />
     <ClInclude Include="..\Enums.h" />
     <ClInclude Include="..\Rendering.h" />
     <ClInclude Include="..\Structs.h" />
